Add routing with React Router
