+++ src/pages/Home.tsx
@@ -1,0 +1,11 @@
+function Home() {
+  return (
+    <div id="home">
+      <h1>Home Page</h1>
+      <p>Welcome!</p>
+    </div>
+  )
+}
+
+export default Home
+
