+++ src/pages/Login.tsx
@@ -1,0 +1,41 @@
+import { useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+
+function Login() {
+  const [username, setUsername] = useState('')
+  const [password, setPassword] = useState('')
+  const navigate = useNavigate()
+
+  const handleLogin = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault()
+    if (username && password) {
+      navigate('/')
+    }
+  }
+
+  return (
+    <form onSubmit={handleLogin}>
+      <h1>Login</h1>
+      <div>
+        <input
+          type="text"
+          placeholder="Username"
+          value={username}
+          onChange={(e) => setUsername(e.target.value)}
+        />
+      </div>
+      <div>
+        <input
+          type="password"
+          placeholder="Password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+        />
+      </div>
+      <button type="submit">Login</button>
+    </form>
+  )
+}
+
+export default Login
+
